+++ packages/webiny-api-cms/index.js
@@ -1,2 +1,1 @@
 // @flow
-export { dataSource } from "./src";
