+++ packages/webiny-api-cms/src/index.js
@@ -1,2 +1,1 @@
 // @flow
-export { default as dataSource } from "./dataSource";
