+++ packages/webiny-api-cms/src/plugins/graphql/Page.js
@@ -6,7 +6,7 @@ import {
     resolveGet,
     resolveList
 } from "webiny-api/graphql";
-import UserType from "webiny-api/dataSource/typeDefs/User";
+import UserType from "webiny-api/plugins/graphql/User";
 import createRevisionFrom from "./pageResolvers/createRevisionFrom";
 import listPages from "./pageResolvers/listPages";
 import listPublishedPages from "./pageResolvers/listPublishedPages";
