+++ packages/webiny-api-cms/src/plugins/graphql/pageResolvers/listPages.js
@@ -8,8 +8,7 @@ type EntityFetcher = (context: Object) => Class<Entity>;
 export default (entityFetcher: EntityFetcher) => async (
     root: any,
     args: Object,
-    context: Object,
-    info: Object
+    context: Object
 ) => {
     const entityClass = entityFetcher(context);
