+++ packages/webiny-api-cms/src/plugins/index.js
@@ -3,4 +3,6 @@ import general from "./pageSettings/general";
 import seo from "./pageSettings/seo";
 import social from "./pageSettings/social";
 import entities from "./entities";
-export default [...general, ...seo, ...social, ...entities];
+import graphql from "./graphql";
+
+export default [...general, ...seo, ...social, ...entities, graphql];
