+++ packages/webiny-api/index.js
@@ -1,2 +1,2 @@
 // @flow
-export { MySQLTable, createHandler, dataSource} from "./src";
+export { MySQLTable, createHandler } from "./src";
