+++ packages/webiny-api/src/index.js
@@ -1,4 +1,8 @@
 // @flow
-export { default as dataSource } from "./dataSource";
+import apiPlugins, { addPlugin } from "./plugins";
+
+// Register required core plugins.
+addPlugin(...apiPlugins);
+
 export { createHandler } from "./lambda/lambda";
 export { default as MySQLTable } from "./mysql";
