+++ packages/webiny-api/src/plugins/graphql/userResolvers/getCurrentUserSettings.js
