+++ packages/webiny-api/src/plugins/graphql/userResolvers/updateCurrentUserSettings.js
